Run 'forge fmt'

+++ src/VeganRestaurant.sol
@@ -14,7 +14,7 @@ contract VeganRestaurant {
 
     modifier onlyOwner(string memory plusCode) {
         require(restaurants[plusCode].owner == msg.sender, "Only the owner can perform this action.");
-        
+
         _;
     }
 
@@ -24,19 +24,21 @@ contract VeganRestaurant {
 
         restaurants[plusCode] = Restaurant({owner: msg.sender, avgRating: 0, totalRatings: 0});
     }
+
     function deleteRestaurant(string memory plusCode) public onlyOwner(plusCode) {
         delete restaurants[plusCode];
     }
+
     function modifyRestaurant(string memory plusCode, string memory newPlusCode) external onlyOwner(plusCode) {
         require(bytes(plusCode).length > 0, "Plus code cannot be empty.");
         require(restaurants[plusCode].owner != address(0), "Restaurant does not exist.");
 
-        uint currentAvgRating = restaurants[plusCode].avgRating;
-        uint currentTotalRatings = restaurants[plusCode].totalRatings;
+        uint256 currentAvgRating = restaurants[plusCode].avgRating;
+        uint256 currentTotalRatings = restaurants[plusCode].totalRatings;
 
         deleteRestaurant(plusCode);
         addRestaurant(newPlusCode);
-        
+
         restaurants[newPlusCode].avgRating = currentAvgRating;
         restaurants[newPlusCode].totalRatings = currentTotalRatings;
     }
